requirements.txt
Django==4.1.7
djangorestframework==3.14.0
djoser==2.1.0
psycopg2-binary==2.9.5
python-environ==0.4.54
drf-yasg==1.21.5
flake8==6.0.0
isort==5.12.0
django_debug_toolbar==3.8.1
gunicorn==20.1.0
Pillow==9.5.0